+++ src/main/java/com/testfabrik/webmate/javasdk/JacksonMapper.java
@@ -1,5 +1,6 @@
 package com.testfabrik.webmate.javasdk;
 
+import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.guava.GuavaModule;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
@@ -20,7 +21,7 @@ public class JacksonMapper {
             JacksonMapper.theInstance.registerModule(new JodaModule());
             JacksonMapper.theInstance.registerModule(new Jdk8Module().configureAbsentsAsNulls(true));
             JacksonMapper.theInstance.registerModule(new GuavaModule());
-
+            JacksonMapper.theInstance.enable(MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS);
         }
         return JacksonMapper.theInstance;
     }